+++ CHANGELOG.md
@@ -16,6 +16,8 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 - [#4679](https://github.com/thanos-io/thanos/pull/4679) Added `enable-feature` flag to enable negative offsets and @ modifier, similar to Prometheus.
 - [#4696](https://github.com/thanos-io/thanos/pull/4696) Query: add cache name to tracing spans.
 - [#4764](https://github.com/thanos-io/thanos/pull/4764) Compactor: add `block-viewer.global.sync-block-timeout` flag to set the timeout of synchronization block metas.
+- [#4389](https://github.com/thanos-io/thanos/pull/4389) Querier: add `endpoint.configuration` and `endpoint.configuration-file` for granular endpoint configuration YAML content or file. 
+  - *:warning:* This also deprecates the following flags `store.sd-interval`, `store.sd-dns-interval`, `store.sd-dns-resolver`, `store.sd-files` and all `grpc-client-.*`. They will be removed in v0.27.0.
 
 ### Fixed